+++ apps/oxlint/test/fixtures/definePlugin_and_defineRule/eslint.config.js
@@ -1,4 +1,4 @@
-import plugin from './plugin.js';
+import plugin from './plugin.ts';
 
 export default [
   {
