+++ apps/oxlint/test/fixtures/defineRule/.oxlintrc.json
@@ -1,5 +1,5 @@
 {
-  "jsPlugins": ["./plugin.js"],
+  "jsPlugins": ["./plugin.ts"],
   "categories": { "correctness": "off" },
   "rules": {
     "define-rule-plugin/create": "error",
